New Login Screen created. Matching figma designs
